Rename: 이전 커밋에 따라 파일 이름 수정
