+++ src/main/kotlin/com/knu/mockin/service/TradingService.kt
@@ -1,11 +1,11 @@
 package com.knu.mockin.service
 
-import com.knu.mockin.kisclient.KISOverSeaClient
+import com.knu.mockin.kisclient.KISTradingClient
 import com.knu.mockin.logging.model.LogAPIEntry
 import com.knu.mockin.logging.utils.LogUtil
 import com.knu.mockin.model.dto.kisheader.request.KISOverSeaRequestHeaderDto
-import com.knu.mockin.model.dto.kisrequest.order.KISOrderRequestDto
-import com.knu.mockin.model.dto.kisresponse.order.KISOverSeaResponseDto
+import com.knu.mockin.model.dto.kisrequest.trading.KISOrderRequestDto
+import com.knu.mockin.model.dto.kisresponse.trading.KISOrderResponseDto
 import com.knu.mockin.model.enum.ExchangeCode
 import com.knu.mockin.model.enum.TradeId
 import com.knu.mockin.repository.UserRepository
@@ -16,11 +16,11 @@ import org.springframework.stereotype.Service
 
 @Service
 class TradingService(
-    private val kisOverSeaClient: KISOverSeaClient,
+    private val kisTradingClient: KISTradingClient,
     private val userRepository: UserRepository
 ) {
     private val log = LoggerFactory.getLogger(TradingService::class.java)
-    suspend fun order(): KISOverSeaResponseDto {
+    suspend fun order(): KISOrderResponseDto {
         val user = userRepository.findById(1).awaitFirst()
         val kisOverSeaRequestHeaderDto = KISOverSeaRequestHeaderDto(
             authorization = "Bearer ${user.token}",
@@ -38,6 +38,10 @@ class TradingService(
         )
         log.info("{}", LogUtil.toJson(LogAPIEntry(kisOverSeaRequestHeaderDto)))
         log.info("{}", LogUtil.toJson(LogAPIEntry(kisOrderRequestDto)))
-        return kisOverSeaClient.postOrder(kisOverSeaRequestHeaderDto, kisOrderRequestDto).awaitSingle()
+        return kisTradingClient.postOrder(kisOverSeaRequestHeaderDto, kisOrderRequestDto).awaitSingle().component2()
+    }
+
+    suspend fun inquireBalance(){
+
     }
 }
